refactor: 함수형 컴포넌트로 변환(jsx) 및 useContext 활용
