+++ src/components/Header.jsx
@@ -2,18 +2,16 @@ import React from "react";
 import styled from "styled-components";
 import addButton from "../assets/images/add-button.png";
 
-class Header extends React.Component {
-  render() {
-    return (
-      <HeaderComponent>
-        <Title>점심 뭐 먹지</Title>
-        <Button onClick={() => alert('지원하지 않는 기능입니다.')}>
-          <ButtonImage src={addButton} alt="음식점 추가" />
-        </Button>
-      </HeaderComponent>
-    )
-  }
-}
+const Header = () => {
+  return (
+    <HeaderComponent>
+      <Title>점심 뭐 먹지</Title>
+      <Button onClick={() => alert("지원하지 않는 기능입니다.")}>
+        <ButtonImage src={addButton} alt="음식점 추가" />
+      </Button>
+    </HeaderComponent>
+  );
+};
 export default Header;
 
 const HeaderComponent = styled.div`
@@ -25,7 +23,7 @@ const HeaderComponent = styled.div`
   padding: 0 16px;
 
   background-color: #ec4a0a;
-`
+`;
 
 const Title = styled.div`
   color: #fcfcfd;
@@ -33,7 +31,7 @@ const Title = styled.div`
   font-size: 20px;
   line-height: 24px;
   font-weight: 600;
-`
+`;
 
 const Button = styled.button`
   height: 40px;
@@ -44,11 +42,11 @@ const Button = styled.button`
 
   font-size: 24px;
   cursor: pointer;
-`
+`;
 
 const ButtonImage = styled.img`
   display: block;
   width: 40px;
   height: 40px;
   object-fit: contain;
-`
+`;
